+++ src/pages/CreateNewGame.tsx
@@ -1,0 +1,14 @@
+import Button from "../components/Button";
+
+function CreateNewGame() {
+
+    return (
+        <>
+
+            <Button buttonTitle="Next" eventHandler={() => { }} />
+        </>
+
+    );
+}
+
+export default CreateNewGame;
