+++ src/pages/HomePage.tsx
@@ -1,0 +1,24 @@
+import Button from "../components/Button";
+
+function HomePage() {
+
+    return (
+        <div>
+            <h1>Game of trivia</h1>
+            <Button buttonTitle="Join Game" eventHandler={() => { }} />
+            <Button buttonTitle="Create Game" eventHandler={() => { }} />
+            <div className="checkbox-container">
+                <div id="checkBox">
+                    <label >
+
+                        <input type="checkbox" />
+                        <span>Horror</span>
+
+                    </label>
+                </div>
+            </div>
+
+        </div>);
+}
+
+export default HomePage;
