--- src/pages/HomePage/HomePage.tsx
@@ -1,15 +1,0 @@
-import Button from "../../components/Button";
-
-function HomePage() {
-
-    return (
-        <div>
-            <h1>Game of trivia</h1>
-            <Button buttonTitle="Join Game" eventHandler={() => { }} />
-            <Button buttonTitle="Create Game" eventHandler={() => { }} />
-
-
-        </div>);
-}
-
-export default HomePage;
